+++ src/app.ts
@@ -2,6 +2,13 @@ import Koa from 'koa'
 import bodyParser from 'koa-bodyparser'
 import router from './routes/router'
 import { errorMiddleware } from './middlewares/error_middleware'
+import type { Repositories } from './respositories'
+
+declare module 'koa' {
+  interface DefaultContext {
+    repositories: Repositories
+  }
+}
 
 const app = new Koa()
 app
