+++ src/models/BoundaryModel.ts
@@ -22,16 +22,9 @@ export interface IBoundary {
   placeId: string
 }
 
-export interface BoundaryModel extends Model<IBoundary> {
-  findBoundaries(
-    lat: number,
-    lng: number,
-    boundaryType: string
-  ): Promise<IBoundary[]>
-}
-
-// Making placeId a required field because we don't really want to have to guess if a boundary matches a given
-// request or not by just searching via `$geoIntersects`. If we were unable to geocode a boundary, or manually, map
+// Making placeId a required field because we don't really want to have to guess
+// if a boundary matches a given request or not by just searching via
+// `$geoIntersects`. If we were unable to geocode a boundary, or manually, map
 // it to a specific string, then we just shouldn't use it at all.
 export const BoundarySchema: Schema<IBoundary> = new Schema({
   name: {
@@ -45,8 +38,9 @@ export const BoundarySchema: Schema<IBoundary> = new Schema({
   },
   geometry: {
     type: MultiPolygonSchema,
-    // NOTE: it's very important that this index gets defined here, rather than on the coordinates in the
-    // MultiPolygonSchema. putting them on the coordinates breaks things so that you can never create a record
+    // NOTE: It's very important that this index gets defined here, rather than
+    // on the coordinates in the MultiPolygonSchema. putting them on the
+    // coordinates breaks things so that you can never create a record
     // successfully
     index: '2dsphere',
     required: true
@@ -58,29 +52,5 @@ export const BoundarySchema: Schema<IBoundary> = new Schema({
   }
 })
 
-BoundarySchema.statics.findBoundaries = async function (
-  lat: number,
-  lng: number,
-  boundaryType: BoundaryType
-) {
-  return this.find({
-    $and: [
-      {
-        geometry: {
-          $geoIntersects: {
-            $geometry: {
-              type: 'Point',
-              coordinates: [lng, lat]
-            }
-          }
-        }
-      },
-      {
-        type: boundaryType
-      }
-    ]
-  })
-}
-
-export default (mongoose.models.Boundary as BoundaryModel) ||
-  mongoose.model<IBoundary, BoundaryModel>('Boundary', BoundarySchema);
+export default (mongoose.models.Boundary as Model<IBoundary>) ||
+  mongoose.model<IBoundary, Model<IBoundary>>('Boundary', BoundarySchema)
