+++ src/respositories/boundaryRepository.ts
@@ -1,0 +1,54 @@
+import BoundaryModel, {
+  type BoundaryType,
+  type IBoundary
+} from '../models/BoundaryModel'
+
+export type BoundaryRepository = {
+  findBoundaries: (
+    lat: number,
+    lng: number,
+    boundaryType: BoundaryType
+  ) => Promise<IBoundary[]>
+
+  findByPlaceId: (placeId: string) => Promise<IBoundary | null>
+
+  findById: (id: string) => Promise<IBoundary | null>
+}
+
+const findBoundaries = async (
+  lat: number,
+  lng: number,
+  boundaryType: BoundaryType
+) => {
+  return BoundaryModel.find({
+    $and: [
+      {
+        geometry: {
+          $geoIntersects: {
+            $geometry: {
+              type: 'Point',
+              coordinates: [lng, lat]
+            }
+          }
+        }
+      },
+      {
+        type: boundaryType
+      }
+    ]
+  })
+}
+
+const findByPlaceId = async (placeId: string) => {
+  return BoundaryModel.findOne({ placeId })
+}
+
+const findById = async (id: string) => {
+  return BoundaryModel.findById(id)
+}
+
+export const boundaryRepository: BoundaryRepository = {
+  findBoundaries,
+  findByPlaceId,
+  findById
+}
