+++ src/server.ts
@@ -1,10 +1,12 @@
 import app from './app'
 import env from './lib/env'
 import { connectToDatabase, disconnectDatabase } from './database'
+import repositories from './respositories'
 
 const startServer = async (): Promise<void> => {
   try {
     await connectToDatabase()
+    app.context.repositories = repositories
     app.on('error', (err, ctx) => {
       console.error('Server error', err, ctx)
     })
