+++ src/test/routes/listingSearchRouter.test.ts
@@ -4,6 +4,7 @@ import Boundary from '../../models/BoundaryModel'
 import Listing from '../../models/ListingModel'
 import type { ListingData } from '../../lib/random_data'
 import type { ListingSearchResponse } from '../../types/listing_search_response_types'
+import repositories from '../../respositories'
 
 const ViewportBoundsFremont = {
   bounds_north: 47.69011227856514,
@@ -51,6 +52,8 @@ const outsideBoundsListingData: ListingData = {
   },
   neighborhood: 'Wallingford'
 }
+
+app.context.repositories = repositories
 
 describe('listingSearchRouter', () => {
   describe('GET /listing/search/boundary/:id', () => {
