+++ src/views/listingSearchBoundaryView.ts
@@ -1,13 +1,12 @@
-import type { ListingSearchAggregateResult } from '../models/ListingModel'
 import type { IBoundary } from '../models/BoundaryModel'
-import type { PaginationParams } from '../zod_schemas/listingSearchParamsSchema'
+import { type FindWithinBoundsResult } from '../respositories/listingRepository'
 import type { BoundarySearchResponse } from '../types/listing_search_response_types'
-import type { ListingResultWithSelectedFields } from '../types/listing_search_response_types'
+import type { PaginationParams } from '../zod_schemas/listingSearchParamsSchema'
 import listingSearchView from './listingSearchView'
 
 export default (
   boundary: IBoundary,
-  results: ListingSearchAggregateResult<ListingResultWithSelectedFields>,
+  results: FindWithinBoundsResult[],
   pagination: PaginationParams
 ): BoundarySearchResponse => {
   return {
